move namespace method definitions to their own cpp file to keep the widget cpp lean (sorry, couldn't come up with a better name...)
